+++ backend/canisters/user_index/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Restart payments job if final payment in queue fails ([#3551](https://github.com/open-ic/open-chat/pull/3551))
+- Append a suffix when registering if username is taken ([#3553](https://github.com/open-ic/open-chat/pull/3553))
 
 ### Fixed